requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
httpx==0.27.0
python-dotenv==1.0.1
redis==5.0.7
django-redis==5.4.0
